Extract shrug, tableflip, unflip, lenny into canned-messages
